refactor: remove the dependency on InputView
